requirements.txt
flask
torch
transformers
Pillow
numpy
scikit-learn